fix Ca value to double

+++ src/ru/proletov/ckdmbd/models/Ca.java
@@ -2,7 +2,7 @@ package ru.proletov.ckdmbd.models;
 
 public class Ca extends LaboratoryInvestigation {
 
-    public Ca(final float value) {
+    public Ca(final double value) {
         super(value);
     }
